Se crea los metodos de CRUD para proyectos + implementacion de sessiones para el modulo de usuarios para dar seguridad a la informacion de los usuarios + se verifica la navegacion entre secciones y sitios web.

+++ src/main/java/com/siesweb/controlador/UsuarioServlet.java
@@ -12,6 +12,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.swing.JOptionPane;
 import org.mindrot.jbcrypt.BCrypt;
 
@@ -19,7 +20,7 @@ import com.siesweb.modelo.*;
 import com.siesweb.dao.*;
 
 @WebServlet({ "/UsuarioServlet", "/iniciarSesion", "/registrar", "/insertar","/actualizar", "/eliminar", "/listar", "/buscar"
-	,"/cargarRegistro","/cargarInsercion","/cargarActualizacion"})
+	,"/cargarRegistro","/cargarInsercion","/cargarActualizacion","/logout","/inactivar"})
 public class UsuarioServlet extends HttpServlet {
 
 	private static final long serialVersionUID = 1L;
@@ -70,6 +71,17 @@ public class UsuarioServlet extends HttpServlet {
 		case "/cargarActualizacion":			
 			loadActualizacion(request, response);
 			break;
+		case "/inactivar":			
+			inactivar(request, response);
+			break;
+		case "/logout":			
+			HttpSession session = request.getSession(false); // No crea una nueva sesión si no existe
+	        if (session != null) {
+	            session.invalidate(); // Invalida la sesión actual
+	        }
+	        RequestDispatcher dispatcher = request.getRequestDispatcher("index.jsp");
+			dispatcher.forward(request, response);
+			break;
 		default:
 			System.out.println("No se reconoce la opcion enviada!");
 		}
@@ -95,7 +107,8 @@ public class UsuarioServlet extends HttpServlet {
 				if (BCrypt.checkpw(pass, passHash)) {
 					int idRol = usuario.getRolId();
 					String nombreUser = usuario.getNombre();
-					request.setAttribute("nombreUser", nombreUser);
+					HttpSession session = request.getSession();			        
+					session.setAttribute("nombreUser", nombreUser);
 					if (idRol == 1) {
 						RequestDispatcher dispatcher = request.getRequestDispatcher("admin.jsp");
 						dispatcher.forward(request, response);
@@ -154,13 +167,14 @@ public class UsuarioServlet extends HttpServlet {
 						&& validarPassword(password) && tipo != "-" && rol != "-" && proyecto != "-") {
 
 					if (tipoDocumento != null && proy != null && rl != null) {
-
+						
 						usuarioDAO.agregarUsuario(new Usuarios(0, nombre, apellido, documento, email, telefono, usuario,
 								hashedPassword, estado, tipoDocumento.getId(), proy.getId(), rl.getId()));
 						JOptionPane.showMessageDialog(null, "Usuario agregado con exito.", "!Advertencia¡",
 								JOptionPane.INFORMATION_MESSAGE);
 						RequestDispatcher dispatcher = request.getRequestDispatcher("Registro.jsp");
 						dispatcher.forward(request, response);
+						
 					} else {
 						System.out.println("No se pudo obtener el id de tipo, rol o proyecto.");
 					}
@@ -215,13 +229,20 @@ public class UsuarioServlet extends HttpServlet {
 						&& validarPassword(password)) {
 
 					if (tipoDocumento != null && proy != null && rl != null) {
-
-						usuarioDAO.agregarUsuario(new Usuarios(0, nombre, apellido, documento, email, telefono, usuario,
-								hashedPassword, estado, tipoDocumento.getId(), proy.getId(), rl.getId()));
-						JOptionPane.showMessageDialog(null, "Usuario agregado con exito.", "!Advertencia¡",
-								JOptionPane.INFORMATION_MESSAGE);
-						RequestDispatcher dispatcher = request.getRequestDispatcher("Registro.jsp");
-						dispatcher.forward(request, response);
+						HttpSession session = request.getSession(false);
+						if (session != null && session.getAttribute("nombreUser") != null) {
+							usuarioDAO.agregarUsuario(new Usuarios(0, nombre, apellido, documento, email, telefono, usuario,
+									hashedPassword, estado, tipoDocumento.getId(), proy.getId(), rl.getId()));
+							JOptionPane.showMessageDialog(null, "Usuario agregado con exito.", "!Advertencia¡",
+									JOptionPane.INFORMATION_MESSAGE);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("adminAgregarUser.jsp");
+							dispatcher.forward(request, response);
+						}else {
+							JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
+									JOptionPane.INFORMATION_MESSAGE);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("indexjsp");
+							dispatcher.forward(request, response);
+						}
 					} else {
 						System.out.println("No se pudo obtener el id de tipo, rol o proyecto.");
 					}
@@ -229,14 +250,14 @@ public class UsuarioServlet extends HttpServlet {
 					System.out.println("Uno o varios campos no son validos.");
 					JOptionPane.showMessageDialog(null, "Uno o varios campos no son validos.", "!Advertencia¡",
 							JOptionPane.INFORMATION_MESSAGE);
-					RequestDispatcher dispatcher = request.getRequestDispatcher("Registro.jsp");
+					RequestDispatcher dispatcher = request.getRequestDispatcher("adminAgregarUser.jsp");
 					dispatcher.forward(request, response);
 				}
 			} else {
 				System.out.println("Debe llenar todos los campos.");
 				JOptionPane.showMessageDialog(null, "Debe llenar todos los campos.", "!Advertencia¡",
 						JOptionPane.INFORMATION_MESSAGE);
-				RequestDispatcher dispatcher = request.getRequestDispatcher("Registro.jsp");
+				RequestDispatcher dispatcher = request.getRequestDispatcher("adminAgregarUser.jsp");
 				dispatcher.forward(request, response);
 			}
 		} catch (SQLException e) {
@@ -279,18 +300,25 @@ public class UsuarioServlet extends HttpServlet {
 						Roles rl = rolesDAO.obtenerIdRol(rol);
 
 						if (tipoDocumento != null && proy != null && rl != null) {
-
-							int respuesta = JOptionPane.showConfirmDialog(null, "¿Estás seguro?");
-							if (respuesta == JOptionPane.YES_OPTION) {
-
-								usuarioDAO.actualizarUsuario(new Usuarios(id, nombre, apellido, documento, email,
-										telefono, usuario, hashedPassword, estado, tipoDocumento.getId(), proy.getId(), rl.getId()));
-								RequestDispatcher dispatcher = request.getRequestDispatcher("adminActualizarUser.jsp");
+							HttpSession session = request.getSession(false);
+							if (session != null && session.getAttribute("nombreUser") != null) {
+								int respuesta = JOptionPane.showConfirmDialog(null, "¿Estás seguro?");
+								if (respuesta == JOptionPane.YES_OPTION) {
+	
+									usuarioDAO.actualizarUsuario(new Usuarios(id, nombre, apellido, documento, email,
+											telefono, usuario, hashedPassword, estado, tipoDocumento.getId(), proy.getId(), rl.getId()));
+									RequestDispatcher dispatcher = request.getRequestDispatcher("adminActualizarUser.jsp");
+									dispatcher.forward(request, response);
+								} else {
+									RequestDispatcher dispatcher = request.getRequestDispatcher("adminActualizarUser.jsp");
+									dispatcher.forward(request, response);
+								}
+							}else {
+								JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
+										JOptionPane.INFORMATION_MESSAGE);
+								RequestDispatcher dispatcher = request.getRequestDispatcher("indexjsp");
 								dispatcher.forward(request, response);
-							} else {
-								RequestDispatcher dispatcher = request.getRequestDispatcher("adminActualizarUser.jsp");
-								dispatcher.forward(request, response);
-							}
+							}	
 						} else {
 							System.out.println("No se pudo obtener el id de tipo, rol o proyecto.");
 						}
@@ -329,13 +357,22 @@ public class UsuarioServlet extends HttpServlet {
 				try {
 					List<Usuarios> user = usuarioDAO.buscarUsuario(id);
 					if (!user.isEmpty()) {
-						int respuesta = JOptionPane.showConfirmDialog(null, "¿Estás seguro?");
-						if (respuesta == JOptionPane.YES_OPTION) {
-							usuarioDAO.eliminarUsuario(id);
-							RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
-							dispatcher.forward(request, response);
-						} else {
-							RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+						HttpSession session = request.getSession(false);
+						if (session != null && session.getAttribute("nombreUser") != null) {
+						
+							int respuesta = JOptionPane.showConfirmDialog(null, "¿Estás seguro?");
+							if (respuesta == JOptionPane.YES_OPTION) {
+								usuarioDAO.eliminarUsuario(id);
+								RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+								dispatcher.forward(request, response);
+							} else {
+								RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+								dispatcher.forward(request, response);
+							}
+						}else {
+							JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
+									JOptionPane.INFORMATION_MESSAGE);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("index.jsp");
 							dispatcher.forward(request, response);
 						}
 					} else {
@@ -365,10 +402,18 @@ public class UsuarioServlet extends HttpServlet {
 
 	private void listar(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
-			List<Usuarios> usuarios = usuarioDAO.obtenerUsuarios();
-			request.setAttribute("listaUsuarios", usuarios);
-			RequestDispatcher dispatcher = request.getRequestDispatcher("adminListarUser.jsp");
-			dispatcher.forward(request, response);
+			HttpSession session = request.getSession(false);
+			if (session != null && session.getAttribute("nombreUser") != null) {
+				List<Usuarios> usuarios = usuarioDAO.obtenerUsuarios();
+				request.setAttribute("listaUsuarios", usuarios);
+				RequestDispatcher dispatcher = request.getRequestDispatcher("adminListarUser.jsp");
+				dispatcher.forward(request, response);
+			}else {
+				JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
+						JOptionPane.INFORMATION_MESSAGE);
+				RequestDispatcher dispatcher = request.getRequestDispatcher("index.jsp");
+				dispatcher.forward(request, response);
+			}
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
@@ -381,17 +426,25 @@ public class UsuarioServlet extends HttpServlet {
 			if (validarId(idStr)) {
 				int id = Integer.parseInt(idStr);
 				try {
-					List<Usuarios> user = usuarioDAO.buscarUsuario(id);
-					if (!user.isEmpty()) {
-
-						request.setAttribute("User", user);
-						RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
-						dispatcher.forward(request, response);
-					} else {
-						System.out.println("Usuario no encontrado.");
-						JOptionPane.showMessageDialog(null, "Usuario no encontrado.", "!Advertencia¡",
+					HttpSession session = request.getSession(false);
+					if (session != null && session.getAttribute("nombreUser") != null) {
+						List<Usuarios> user = usuarioDAO.buscarUsuario(id);
+						if (!user.isEmpty()) {
+	
+							request.setAttribute("User", user);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+							dispatcher.forward(request, response);
+						} else {
+							System.out.println("Usuario no encontrado.");
+							JOptionPane.showMessageDialog(null, "Usuario no encontrado.", "!Advertencia¡",
+									JOptionPane.INFORMATION_MESSAGE);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+							dispatcher.forward(request, response);
+						}
+					}else {
+						JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
 								JOptionPane.INFORMATION_MESSAGE);
-						RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+						RequestDispatcher dispatcher = request.getRequestDispatcher("index.jsp");
 						dispatcher.forward(request, response);
 					}
 				} catch (SQLException e) {
@@ -457,6 +510,59 @@ public class UsuarioServlet extends HttpServlet {
         	System.out.println("no se agrego nada");
         }
     }
+	
+	private void inactivar(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+		String idString = request.getParameter("idInactivar");
+		String estado = "Inactivo";
+		if (!idString.isEmpty()) {
+			if (validarId(idString)) {
+				int id = Integer.parseInt(idString);
+				try {
+					HttpSession session = request.getSession(false);
+					if (session != null && session.getAttribute("nombreUser") != null) {
+						List<Usuarios> usuario = usuarioDAO.buscarUsuario(id);
+						
+						if (!usuario.isEmpty()) {
+							int respuesta = JOptionPane.showConfirmDialog(null, "¿Estás seguro?");
+							if (respuesta == JOptionPane.YES_OPTION) {
+								usuarioDAO.inactivarUsuario(new Usuarios(id,"","","","","","","",estado,0,0,0));
+								RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+								dispatcher.forward(request, response);
+							} else {
+								RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+								dispatcher.forward(request, response);
+							}
+						} else {
+							System.out.println("No se encontro el id a inactivar.");
+							JOptionPane.showMessageDialog(null, "No se encontro el id a inactivar.", "!Advertencia¡",
+									JOptionPane.INFORMATION_MESSAGE);
+							RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+							dispatcher.forward(request, response);
+						}
+					}else {
+						JOptionPane.showMessageDialog(null, "Debes volver a iniciar sesión.", "!Advertencia¡",
+								JOptionPane.INFORMATION_MESSAGE);
+						RequestDispatcher dispatcher = request.getRequestDispatcher("index.jsp");
+						dispatcher.forward(request, response);
+					}
+				} catch (SQLException e) {
+				}
+			} else {
+				System.out.println("El id no es valido.");
+				JOptionPane.showMessageDialog(null, "El id no es valido.", "!Advertencia¡",
+						JOptionPane.INFORMATION_MESSAGE);
+				RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+				dispatcher.forward(request, response);
+			}
+		} else {
+			System.out.println("Ingresa un id porfavor");
+			JOptionPane.showMessageDialog(null, "Debes ingresar el id para inactivar.", "!Advertencia¡",
+					JOptionPane.INFORMATION_MESSAGE);
+			RequestDispatcher dispatcher = request.getRequestDispatcher("adminBuscarEliminarUser.jsp");
+			dispatcher.forward(request, response);
+		}
+	}
 
 	public static boolean validarId(String id) {
 		String regex = "^[0-9]{1,3}$";
